Migrate to Biome and None-LS
I have updated LSP configurations such that Biome is now my preferred
Linter of choice for JS projects AND my formatter of choice. It
therefore replaces eslint and prettier. I have also swapped from null_ls
(which was archived) to none_ls (which is a direct fork I think, and is
actively maintained).