Add filetype-specific format-on-save AutoCmd
I have been annoyed that the formatting applied by certain LSPs (or
`none-ls` as the backup) have not matched the (non-existent) formatting
standards for Amazon code packages.

As such, I need a way of programatically specifying filetypes that I do
or don't want format-on-save enabled on.

The first victim? **Smithy**