chore: Reorder nvimcmp sources to prioritise snippets
I am going to experiment with this at work and see if I prefer it to
having the LSP sources come first. I think realistically, I probably
will revert this because my snippets should be done in such a way that
they can automatically trigger and have triggers that don't overlap with
common LSP commands.